Support bundled NakedJSX fallback
This tool will now try and detect whether the input files are part of a node project that imports @nakedjsx/core. If they are, then the npx nakedjsx command will be forwarded to that installation of NakedJSX. If not, then a version of NakedJSX bundled with the tool will operate on the files.